feat: add form to main page and refactor - delete unnecessary method and move interface declaration
