--- src/constants/interfaces.ts
@@ -1,7 +1,0 @@
-export interface ChatItemPreview {
-  name: string;
-  message: string;
-  time: string;
-  unread: string;
-  onChatItemClick?: () => void;
-}
